satisfies all requirements
