+++ intsureview_fe/src/App.css
@@ -1,5 +1,6 @@
 .App {
   text-align: center;
+  background-color: #ddcbb7;
 }
 
 .App-logo {
