+++ intsureview_fe/src/App.js
@@ -1,5 +1,5 @@
 import './App.css';
-import NewForm from './components/NewForm';
+import Form from './components/Form';
 import Header from './components/Header';
 import Footer from './components/Footer';
 
@@ -7,7 +7,7 @@ function App() {
   return (
     <div className="App">
       <Header />
-      <NewForm />
+      <Form />
       <Footer />
     </div>
   );
