+++ intsureview_fe/src/components/Form.js
@@ -1,129 +1,120 @@
-// import Select from 'react-select';
-import { useForm, Controller } from 'react-hook-form';
-// import Input from '@material-ui/core/Input';
-import { Input, Select, Button } from '@mui/material';
-import { useEffect, useState } from 'react';
+import {
+  Select,
+  Button,
+  FormLabel,
+  TextField,
+  MenuItem,
+  Container,
+} from '@mui/material';
+import { useState } from 'react';
 import '../css/Form.css';
 
 const Form = () => {
-  const [data, setData] = useState();
-  const [onSale, setOnSale] = useState(null);
-  const { control, handleSubmit, field } = useForm({
-    defaultValues: {
-      datePurchased: '',
-      onSale: '',
-      price: '',
-      store: '',
-      ounces: '',
-      zipCode: '',
-    },
-  });
+  const [datePurchased, setDatePurchased] = useState('');
+  const [onSale, setOnSale] = useState(false);
+  const [price, setPrice] = useState('');
+  const [ounces, setOunces] = useState(undefined);
+  const [zipCode, setZipCode] = useState('');
 
-  const onSubmit = async (data) => {
-    //   const formData = `{
-    //     "date_purchased": ${data.datePurchased},
-    //     "on_sale": ${data.onSale},
-    //     "price": ${data.price},
-    //     "store_name": ${data.store},
-    //     "ounces_per_week": ${data.ounces},
-    //     "zip_code": ${data.zipCode}
-    // }`;
-
+  const submitSurvey = async () => {
     const formData = `{
-    "date_purchased": "2023-06-23",
-    "on_sale": false,
-    "price": "${data.price}",
-    "store_name": "${data.store}",
-    "ounces_per_week": ${data.ounces},
-    "zip_code": ${data.zipCode}
+        "date_purchased": "${datePurchased}",
+        "on_sale": ${onSale},
+        "price": "${price}",
+        "ounces_per_week": ${ounces},
+        "zip_code": ${zipCode}
     }`;
 
-    console.log(formData);
+    if (isZipCodeValid(zipCode) === false) {
+      return;
+    }
+
     try {
-      const response = await fetch('oatmilk/', {
+      const response = await fetch('survey/', {
         method: 'POST',
         headers: {
           'Content-Type': 'application/json',
         },
         body: formData,
       });
-      // setData(response);
+      alert(response.statusText);
     } catch (e) {
       console.error(e);
     }
   };
 
-  let options1 = [
-    { value: 'false', label: 'no' },
-    { value: 'true', label: 'yes' },
-  ];
+  const isZipCodeValid = (zipCode) => {
+    if (zipCode.length === 5 && zipCode.match(/^[0-9]+$/)) {
+      return true;
+    } else {
+      alert('Enter a valid zip code!');
+      return false;
+    }
+  };
 
   return (
-    <form onSubmit={handleSubmit(onSubmit)} className="form">
-      When was the last time you purchased oat milk?
-      <br />
-      <Controller
-        name="datePurchased"
-        control={control}
-        render={({ field }) => <Input label="YYYY-DD-MM" {...field} />}
-      />
-      <br />
-      Was it on sale?
-      <br />
-      {/* <Controller
-        name="onSale"
-        control={control}
-        render={({ ...field }) => (
-          <Select
-            {...field}
-            options={options1}
-            value={options1.find((e) => e.value === onSale)}
+    <Container>
+      <form>
+        <div>
+          <FormLabel className="form-label">
+            When was the last time you purchased oat milk?
+          </FormLabel>
+          <TextField
+            className="text-field"
+            required={true}
+            label="YYYY-MM-DD"
+            onChange={(e) => setDatePurchased(e.target.value)}
+            value={datePurchased}
           />
-        )}
-      /> */}
-      <br />
-      How much did it cost?
-      <br />
-      <Controller
-        name="price"
-        control={control}
-        render={({ field }) => <Input {...field} />}
-      />
-      <br />
-      What is the name of the store you purchased it from?
-      <br />
-      <Controller
-        name="store"
-        control={control}
-        render={({ field }) => <Input {...field} />}
-      />
-      <br />
-      How many ounces of oat milk do you drink per week?
-      <br />
-      {/* this can be a number counter going up and down */}
-      <Controller
-        name="ounces"
-        control={control}
-        render={({ field }) => <Input {...field} />}
-      />
-      <br />
-      What is your zip code?
-      <br />
-      <Controller
-        name="zipCode"
-        control={control}
-        render={({ field }) => <Input {...field} />}
-      />
-      <br />
-      <Button variant="contained" type="submit">
-        Submit
-      </Button>
-      {/* {data ? (
-        <div>Thank you for your entry!</div>
-      ) : (
-        <div>Something went wrong!</div>
-      )} */}
-    </form>
+        </div>
+        <div>
+          <FormLabel className="form-label">Was it on sale?</FormLabel>
+          <Select
+            className="select"
+            onChange={(e) => setOnSale(e.target.value)}
+            value={onSale}
+          >
+            <MenuItem value={true}>Yes</MenuItem>
+            <MenuItem value={false}>No</MenuItem>
+          </Select>
+        </div>
+        <div>
+          <FormLabel className="form-label">How much did it cost?</FormLabel>
+          <TextField
+            className="text-field"
+            required={true}
+            label="$"
+            onChange={(e) => setPrice(e.target.value)}
+            value={price}
+          />
+        </div>
+        <div>
+          <FormLabel className="form-label">
+            How much oat milk do you drink per week?
+          </FormLabel>
+          <TextField
+            className="text-field"
+            required={true}
+            label="ounces"
+            onChange={(e) => setOunces(e.target.value)}
+            value={ounces}
+          />
+        </div>
+        <div>
+          <FormLabel className="form-label">What is your zip code?</FormLabel>
+          <TextField
+            className="text-field"
+            required={true}
+            label="zip code"
+            onChange={(e) => setZipCode(e.target.value)}
+            value={zipCode}
+          />
+        </div>
+        <Button variant="contained" type="button" onClick={submitSurvey}>
+          Submit
+        </Button>
+      </form>
+    </Container>
   );
 };
